+++ tests/Silex/Tests/Application/MailerTraitTest.php
@@ -12,19 +12,19 @@
 namespace Silex\Tests\Application;
 
 use PHPUnit\Framework\TestCase;
-use Silex\Provider\SwiftmailerServiceProvider;
+use Silex\Provider\MailerServiceProvider;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class SwiftmailerTraitTest extends TestCase
+class MailerTraitTest extends TestCase
 {
     public function testMail()
     {
         $app = $this->createApplication();
 
-        $message = $this->getMockBuilder('Swift_Message')->disableOriginalConstructor()->getMock();
-        $app['mailer'] = $mailer = $this->getMockBuilder('Swift_Mailer')->disableOriginalConstructor()->getMock();
+        $message = $this->getMockBuilder('Message')->disableOriginalConstructor()->getMock();
+        $app['mailer'] = $mailer = $this->getMockBuilder('Mailer')->disableOriginalConstructor()->getMock();
         $mailer->expects($this->once())
                ->method('send')
                ->with($message)
@@ -35,8 +35,8 @@ class SwiftmailerTraitTest extends TestCase
 
     public function createApplication()
     {
-        $app = new SwiftmailerApplication();
-        $app->register(new SwiftmailerServiceProvider());
+        $app = new MailerApplication();
+        $app->register(new MailerServiceProvider());
 
         return $app;
     }